RegTraka Backend/pyproject.toml
[tool.poetry]
name = "regtraka"
version = "0.1.0"
description = ""
authors = ["AbdulAfeez <abdulafeez.adeyemo@kibo.school>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.0.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.2.0"
opencv-python = "^4.8.0.76"
face-recognition = "^1.3.0"
django-filter = "^23.3"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
